feat: Add XP caption button theme for non-native titlebar
WLM used a slightly different caption button theme on Windows XP compared
to on Vista+. The caption buttons have a look similar to older Chrome
versions with a transparent close button and a flatter look.

The XP caption buttons are added as an option that the user can enable in
the settings. It is not enabled by default.

Additionally, this commit may optimise memory usage of the caption button
images globally, as each image is only loaded once rather than per-window.

Signed-off-by: Isabella <kawapure@gmail.com>