bugfix+feat: fixes Lycan, luisl bugs and general complaints
I've added a close button to the image viewer as well as glow effects to the button to fit the aero aesthetic better. I've also added tooltips to help viewers understand what the controls do. I've disabled the opening animation in order to fix the Lycan and luisl bugs and also disabled the close on sizing to allow the user to resize to their heart's content. Added a converter to the recipient's profile picture frame to avoid exceptions when not in a DM which marginally improves performance.