feat: Proper system menu handling with custom window frames
Previously, the system menu was not opened if you clicked the application
icon in the titlebar, but there was support for the double-click-to-close
behaviour of the button.

This commit modifies the window procedure hook to handle this element
properly. Now the native system menu is opened when you click the item.

Additionally, some of the BasicTitlebar.cs code was refactored.

Signed-off-by: Isabella <kawapure@gmail.com>