feat: Improve popup behaviour
Popups certainly feel better after this update. They are no longer forced
on top of all windows on the system. They now move alongside the parent
window, quite smoothly even. They now even close when clicked off of.

There are still a few bugs as of the time of committing, but it certainly
isn't terrible.

Signed-off-by: Isabella <kawapure@gmail.com>