fix: Better error handling when chats are unavailable (#47)
Previously, a server or chat request failing before the chat window became
visible would result in the entire application dying. The error handling is
still bad, but this adds a user-friendly bandaid on it. Now, when a request
fails, the application will cancel the creation of objects pertaining to
the chat view, and a user-friendly message will be displayed to indicate
the error state.

Signed-off-by: Isabella <kawapure@gmail.com>