feat(AttachmentEditor): Anchor popup to scroll viewer
This commit changes the popup to be anchored to the scroll viewer instead
of the item which opened it. This means that the popup is now stationary
relative to the parent window. I think I prefer this a lot more.

The only concern that I have now is that flickering is especially visible
during switching after this commit. I will come up with some way to rectify
this visual issue in the future.

Signed-off-by: Isabella Lulamoon <kawapure@gmail.com>