feat: 0.0.2.0 - the megacommit!
This commit features all all new additions and features I wrote over the
past couple days. This includes:

- Making Aerochat read the startup status from Discord's user settings,
  as well as synchronising the user state to Discord's user settings. This
  change ensures that your status will be synchronised properly with other
  Discord clients.
- Message context menus.
- The ability to edit messages.
- The ability to reply to messages.
- The ability to delete messages.
- Granular controls for notifications.
- The ability to open notifications windows automatically.
- Options to hide various Aerochat-specific branding features.

This prepares for a new update for Aerochat, version 0.0.2. I bumped the
version number finally due to the massive number of incoming features.

This commit is not even the end. There are some bugs that still need to be
resolved, and some features I would still like to add in before the next
release.

Signed-off-by: Isabella Lulamoon <kawapure@gmail.com>