fix: Title not updating on non-native titlebar
This was due to a faulty WM_SETTEXT handler that went unnoticed this whole
time. The new window title is conveyed through the LPARAM, not the WPARAM.

Additionally, a fallback to query the title directly from the WPF window
context has been added, which also seems to work as a reliable fallback.

Signed-off-by: Isabella Lulamoon <kawapure@gmail.com>