feat: Crash report window on application crash
This is a very useful feature that I wanted to get out before the v0.2 gold
release because it is very helpful for the average user to get immediate
feedback about a crash rather than having to go the event viewer everytime.

Most users don't know about the event viewer.

Signed-off-by: Isabella Lulamoon <kawapure@gmail.com>