fix: #75
The unloaded handler was doing a lot of work that it really didn't need to
do. I think that a lot of that work should have been done on destruction
(which RAII principle implies it does) rather than removal from the visual
tree (in which case it's usually reinserted very shortly after).

An easy way to have triggered this bug, by the way, was to change DPI and
wiggle your mouse over the window during the refresh period.

Signed-off-by: Isabella Lulamoon <kawapure@gmail.com>