fix: Revert changes to clipboard attachments from 1df72c0
I'm not sure what the point of this change was, but it broke pasting at
least most types of images into the client. The original code, which I
restored in this commit, worked just fine.

If the code was to fix a particular edge case, then it will need further
investigation, but as it stands, the current approach will stick.

Signed-off-by: Isabella Lulamoon <kawapure@gmail.com>