feat: Rewrite custom titlebar to not use WPF window chrome
As nice as it would be to rely on native features of the library, if I've
learnt anything in my time working on Aerochat, it's that I (yes, me,
Isabella!) write better code than any of the minds at Microsoft who wrote
the Windows Presentation Foundation.

At that, the built-in WindowChrome functionality is one of the worst parts
of WPF. It has already been responsible for one issue in Aerochat, for
which the monkey patch to solve has been removed in this very commit since
it's no longer needed after replacing WindowChrome. But if this were just a
one-off problem, I wouldn't have bothered rushing to replace it. Not only
did WindowChrome cause crashing which we needed to patch the framework to
work around, but it also can just cause the program to deadlock. If you
resized a window a lot, it'd be really easy to replicate this. As far as I
could tell (guess), this had to do with how they inefficiently reallocate
objects pertaining to the window region (which is used for the corner
mask). My implementation does not have the same problem.

Signed-off-by: Isabella Lulamoon <kawapure@gmail.com>