feat: Dynamic migration, exception handling, and user agent fun
The title, for once, actually says it all. This commit handles the client-
side part of the dynamic content migration from fbf8897. It also adds some
better exception handling around network request blocks because I noticed
that it could be problematic, and I decided to also the code that reports
our user agent so that it matches the proper capitalisation of the project
name and reports the version.

Apart from that, I am quite proud that I managed to actually summarise all
changes made in this commit in the commit title alone, which is really hard
to do given the character limit.

Signed-off-by: Isabella Lulamoon <kawapure@gmail.com>