chore: Shorten lengths of some news bodies
Currently, the client does not scroll long news bodies, so they will get
cut off. This functionality can be implemented into the client itself in
the future, but for now, we have a character limit for news bodies in order
to ensure they display appropriately at the default size of the home
window.

Signed-off-by: Isabella Lulamoon <kawapure@gmail.com>