feat: Update custom apphost + improve building
The custom apphost now uses the "embedded resource" build action, which
makes MSBuild rebuild the application when it is updated.

The custom apphost now enables a mitigation for WriteXorExecute bugginess
in .NET 8, which forces it to be disabled in ALL situations. Depending on
the results of my future testing, I may enable it for later versions of
Windows 10 and Windows 11, while keeping it disabled for Windows 7 and 8.x.

The custom apphost also has been downgraded to the .NET 8 runtime codebase,
which fixes a problem where it would conflict with the .NET 9 runtime
environment if it were installed on the user's system and prevented
application startup. I hadn't looked too far into the issue at hand, but my
guess is that it was not supposed to load the .NET 9 hostfxr library (which
it was) and that the .NET 8 hostfxr library, which it was supposed to use
instead, ran some extra code to set up security features. The .NET 8
coreclr library was still loaded, and was failing at an assertion:

   void Thread::StaticInitialize()
   {
       WRAPPER_NO_CONTRACT;

   #ifdef FEATURE_SPECIAL_USER_MODE_APC
       InitializeSpecialUserModeApc();

       // ...
       _ASSERTE_ALL_BUILDS(!AreCetShadowStacksEnabled() || UseSpecialUserModeApc());
       // ^^ Assertion fail here.
   #endif
   }

This seems to be perfectly mitigated with the downgrade, so I don't think
it is an issue worth looking into, but I wanted to document it nonetheless.

Signed-off-by: Isabella Lulamoon <kawapure@gmail.com>